chore: split results modules
