+++ src/gem/extractors/draft.py
@@ -23,8 +23,8 @@ from gem.constants import HEROES
 from gem.state.entities import Entity, EntityOp
 
 if TYPE_CHECKING:
-    from gem.models import ParsedPlayer
     from gem.parser import ReplayParser
+    from gem.results.models import ParsedPlayer
 
 # Build reverse mapping: hero_id (int) → npc_name (str) from bundled data.
 # This only covers heroes present in heroes.json (IDs ≤ 155 as of the bundled
