+++ src/gem/results/__init__.py
@@ -1,5 +1,7 @@
-"""Compatibility shim for :mod:`gem.results.models`."""
+"""Parsed match result models, assembly, and tabular projections."""
 
+from gem.results.assembly import build_parsed_match
+from gem.results.dataframes import build_dataframes
 from gem.results.models import (
     ChatEntry,
     NeutralItemFoundEvent,
@@ -16,4 +18,6 @@ __all__ = [
     "ParsedPlayer",
     "SmokeEvent",
     "VisionModifierEvent",
+    "build_dataframes",
+    "build_parsed_match",
 ]
