+++ src/gem/rosh_conversion.py
@@ -14,7 +14,7 @@ from typing import TYPE_CHECKING, Literal
 if TYPE_CHECKING:
     from gem.extractors.objectives import AegisEvent
     from gem.extractors.teamfights import Teamfight
-    from gem.models import ParsedMatch
+    from gem.results.models import ParsedMatch
 
 _TEAM_RADIANT = 2
 _TEAM_DIRE = 3
