+++ tests/test_audit_camp_annotations.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from gem.combat.log import CombatLogEntry
-from gem.models import ParsedMatch, ParsedPlayer
+from gem.results.models import ParsedMatch, ParsedPlayer
 from scripts.audit_camp_annotations import (
     CampZone,
     audit_match,
