+++ tests/test_map_context.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from gem.extractors.objectives import AegisEvent, RoshanKill, TowerKill
 from gem.extractors.wards import WardEvent
 from gem.map_context import MapContextBucket, build_map_context_timeline, score_camp_visit_context
-from gem.models import ParsedMatch
+from gem.results.models import ParsedMatch
 
 
 def _make_match() -> ParsedMatch:
